backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Basic utilities
aiofiles==23.2.1
httpx==0.24.1

# LLM and TTS dependencies (now required for manual assistance)
Pillow==10.1.0
requests==2.31.0
# openai==1.51.0  # Using requests for ChatGPT API
# azure-cognitiveservices-speech==1.34.0  # Using requests for Azure TTS

# HMM and scientific computing dependencies
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
# opencv-python-headless>=4.8.0.76

# Eye tracking
tobii-research>=1.10.0  # Tobii Pro SDK for Python
